Always reinitialise variables on reset.
While the Mega Drive does technically allow games to be soft-reset
without resetting any variables, this does not mesh very well with C
and especially C++ (due to destructors being bypassed).

I will add a special section for variables which should not be
affected by soft-resets, for things like recording that the checksum
has already been completed. That should be a good compromise.